# ENH: Avoid use of item XINCREF and DECREF in fasttake