Rather, use the cast function directly when the copy is not trivial (which we know based on it not passing REFCHK, if that passes we assume memcpy is fine).

Also uses memcpy, since no overlap is possible here.


(There is a tests that covers these paths.)

---

This is a followup (in a sense) to the reference counting cleanup.  The functions `PyArray_Item_INCREF` and `PyArray_Item_XDECREF` should be considered deprecated (or plain wrong) now.

Thus, here is a blueprint of now to fix fasttake (and others would have to follow up on that).  `copyswap` would be another solution, but I want to see if we can't get away with this (here we can, its much faster, but it is plausible that copyswap would be even faster, since it is a single item).

---

Benchmark result, just to proof that the old stuff was anyway ridiculous for structured arrays:

```
       before           after         ratio
     [af9f6568]       [38d52100]
     <main>           <fast-take-no-refcnt>
-        8.03±0μs      5.00±0.01μs     0.62  bench_itemselection.Take.time_contiguous((1000, 1), 'wrap', 'O')
-     8.03±0.02μs      5.00±0.06μs     0.62  bench_itemselection.Take.time_contiguous((1000, 1), 'clip', 'O')
-     15.7±0.01μs       9.67±0.1μs     0.61  bench_itemselection.Take.time_contiguous((2, 1000, 1), 'raise', 'O')
-     15.7±0.01μs      9.62±0.01μs     0.61  bench_itemselection.Take.time_contiguous((2, 1000, 1), 'wrap', 'O')
-     15.7±0.01μs      9.46±0.01μs     0.60  bench_itemselection.Take.time_contiguous((2, 1000, 1), 'clip', 'O')
-      8.29±0.3μs      4.97±0.04μs     0.60  bench_itemselection.Take.time_contiguous((1000, 1), 'raise', 'O')
-     15.1±0.02μs      8.70±0.08μs     0.58  bench_itemselection.Take.time_contiguous((1000, 2), 'wrap', 'O')
-     22.2±0.04μs         12.8±0μs     0.58  bench_itemselection.Take.time_contiguous((1000, 3), 'raise', 'O')
-     22.1±0.04μs      12.7±0.01μs     0.58  bench_itemselection.Take.time_contiguous((1000, 3), 'wrap', 'O')
-     22.2±0.02μs      12.7±0.03μs     0.58  bench_itemselection.Take.time_contiguous((1000, 3), 'clip', 'O')
-     15.1±0.01μs      8.67±0.02μs     0.57  bench_itemselection.Take.time_contiguous((1000, 2), 'clip', 'O')
-     15.1±0.01μs      8.65±0.02μs     0.57  bench_itemselection.Take.time_contiguous((1000, 2), 'raise', 'O')
-       177±0.4μs      12.9±0.02μs     0.07  bench_itemselection.Take.time_contiguous((1000, 1), 'wrap', 'i,O')
-         178±1μs       12.9±0.2μs     0.07  bench_itemselection.Take.time_contiguous((1000, 1), 'clip', 'i,O')
-         353±1μs       25.4±0.2μs     0.07  bench_itemselection.Take.time_contiguous((2, 1000, 1), 'raise', 'i,O')
-       189±0.7μs       13.6±0.4μs     0.07  bench_itemselection.Take.time_contiguous((1000, 1), 'raise', 'i,O')
-         355±1μs      25.0±0.04μs     0.07  bench_itemselection.Take.time_contiguous((2, 1000, 1), 'wrap', 'i,O')
-       356±0.4μs      25.0±0.04μs     0.07  bench_itemselection.Take.time_contiguous((2, 1000, 1), 'clip', 'i,O')
-         352±2μs       20.9±0.3μs     0.06  bench_itemselection.Take.time_contiguous((1000, 2), 'wrap', 'i,O')
-       356±0.6μs      20.9±0.07μs     0.06  bench_itemselection.Take.time_contiguous((1000, 2), 'clip', 'i,O')
-         357±3μs      20.9±0.03μs     0.06  bench_itemselection.Take.time_contiguous((1000, 2), 'raise', 'i,O')
-         528±2μs       27.7±0.2μs     0.05  bench_itemselection.Take.time_contiguous((1000, 3), 'wrap', 'i,O')
-         533±2μs      27.6±0.07μs     0.05  bench_itemselection.Take.time_contiguous((1000, 3), 'clip', 'i,O')
-         532±2μs      27.5±0.04μs     0.05  bench_itemselection.Take.time_contiguous((1000, 3), 'raise', 'i,O')
```