@seberg 2023-02-20
In my opinion, in theory the best solution would be to generalize the new `ufunc.at` inner-loop to be usable here (or in a very similar manner).  But it is not there yet, and this is fine.