@seberg 2023-02-20
> Nit: maybe move the NPY_END_THREADS to after the fail: label?

That was intentional, because the index failure already releases it.  OTOH, I guess it should be safe to reset it twice (I am not 100% sure).